MessageLoaderモジュールを削除してsebastian.rbに組み込む
