+++ lib/ruboty/sebastian.rb
@@ -1,5 +1,4 @@
 require "ruboty/sebastian/version"
-require "ruboty/message_loader"
 require "ruboty/actions/greeting_at_morning"
 require "ruboty/actions/greeting_before_closing"
 require "ruboty/actions/greeting_at_closing"
@@ -7,6 +6,13 @@ require "ruboty/handlers/sebastian"
 
 module Ruboty
   module Sebastian
-    # Your code goes here...
+    class Settings < Settingslogic
+      source File.expand_path('../../settings.yml', File.dirname(__FILE__))
+      namespace "messages"
+    end
+
+    def load_message(type)
+      Settings.send(type)
+    end
   end
 end
